fix coolprop test collection collision
Agent-Logs-Url: https://github.com/GBOGEB/CODEX/sessions/3290ddf5-aa99-4166-abc2-01c59af1fcf2
